requirements.txt
 psycopg2-binary
 djangorestframework
 djangorestframework-simplejwt
 django gunicorn
 django-cors-headers


